Reorganized the module and renamed some things
