+++ pygama/dsp/wfs.bin
@@ -1,0 +1,1 @@
+../../base/wfs.bin
